refactored to threadpool class, need to update readme + create singleton class
