+++ GameEngine/src/rendering/Texture.cpp
@@ -5,8 +5,6 @@
 #define STB_IMAGE_IMPLEMENTATION
 #include <stb_image.h>
 
-#include "../Common.h"
-
 void Texture::Cleanup(VkDevice device) {
     vkDestroySampler(device, textureSampler, nullptr);
     vkDestroyImageView(device, textureImageView, nullptr);
